about half way done
