+++ DearPyGui/CMakeLists.txt
@@ -48,6 +48,8 @@ set(MARVEL_SOURCES
 	"src/Core/PythonCommands/mvInterfaceCore.cpp"
 	"src/Core/PythonCommands/mvPlotInterface.cpp"
 	"src/Core/PythonCommands/mvDrawingInterface.cpp"
+	"src/Core/PythonCommands/mvTableInterface.cpp"
+	"src/Core/PythonCommands/mvThemeInterface.cpp"
 
 	"$<$<PLATFORM_ID:Windows>:src/Core/mvMarvel.cpp>"
 	"$<$<PLATFORM_ID:Linux>:src/Core/mvMarvel.cpp>"
