+++ DearPyGui/include/mvMarvel.h
@@ -597,86 +597,6 @@ namespace Marvel {
 		}, "Adds drag for a 4 int values. CTRL+Click to directly modify the value.", "None", "Adding Widgets") });
 	}
 
-	static void AddTableCommands(std::map<std::string, mvPythonParser>* parsers)
-	{
-		parsers->insert({ "add_table", mvPythonParser({
-			{mvPythonDataType::String, "name"},
-			{mvPythonDataType::StringList, "headers"},
-			{mvPythonDataType::KeywordOnly},
-			{mvPythonDataType::Object, "callback", "Registers a callback"},
-			{mvPythonDataType::Object, "callback_data", "Callback data"},
-			{mvPythonDataType::String, "parent", "Parent this item will be added to. (runtime adding)"},
-			{mvPythonDataType::String, "before","This item will be displayed before the specified item in the parent. (runtime adding)"}
-		}, "Adds table.", "None", "Tables") });
-
-		parsers->insert({ "set_table_item", mvPythonParser({
-			{mvPythonDataType::String, "table"},
-			{mvPythonDataType::Integer, "row"},
-			{mvPythonDataType::Integer, "column"},
-			{mvPythonDataType::String, "value"},
-		}, "Sets a table's cell value.", "None", "Tables") });
-
-		parsers->insert({ "delete_row", mvPythonParser({
-			{mvPythonDataType::String, "table"},
-			{mvPythonDataType::Integer, "row"}
-		}, "Delete a row in a table.", "None", "Tables") });
-
-		parsers->insert({ "delete_column", mvPythonParser({
-			{mvPythonDataType::String, "table"},
-			{mvPythonDataType::Integer, "column"}
-		}, "Delete a column in a table.", "None", "Tables") });
-
-		parsers->insert({ "set_headers", mvPythonParser({
-			{mvPythonDataType::String, "table"},
-			{mvPythonDataType::StringList, "headers"},
-		}, "Sets a tables headers.", "None", "Tables") });
-
-		parsers->insert({ "add_row", mvPythonParser({
-			{mvPythonDataType::String, "table"},
-			{mvPythonDataType::StringList, "row"},
-		}, "Adds a row to the end of a table.", "None", "Tables") });
-
-		parsers->insert({ "clear_table", mvPythonParser({
-			{mvPythonDataType::String, "table"}
-		}, "Clears data in a table", "None", "Tables") });
-
-		parsers->insert({ "add_column", mvPythonParser({
-			{mvPythonDataType::String, "table"},
-			{mvPythonDataType::String, "name"},
-			{mvPythonDataType::StringList, "column"},
-		}, "Adds a column to the end of a table.", "None", "Tables") });
-
-		parsers->insert({ "insert_column", mvPythonParser({
-			{mvPythonDataType::String, "table"},
-			{mvPythonDataType::Integer, "column_index"},
-			{mvPythonDataType::String, "name"},
-			{mvPythonDataType::StringList, "column"},
-		}, "Inserts a column into a table.", "None", "Tables") });
-
-		parsers->insert({ "insert_row", mvPythonParser({
-			{mvPythonDataType::String, "table"},
-			{mvPythonDataType::Integer, "row_index"},
-			{mvPythonDataType::StringList, "row"},
-		}, "Inserts a row into a table.", "None", "Tables") });
-
-		parsers->insert({ "set_table_selection", mvPythonParser({
-			{mvPythonDataType::String, "table"},
-			{mvPythonDataType::Integer, "row"},
-			{mvPythonDataType::Integer, "column"},
-			{mvPythonDataType::Bool, "value"},
-		}, "Sets a table's cell selection value.", "None", "Tables") });
-
-		parsers->insert({ "get_table_item", mvPythonParser({
-			{mvPythonDataType::String, "table"},
-			{mvPythonDataType::Integer, "row"},
-			{mvPythonDataType::Integer, "column"}
-		}, "Gets a table's cell value.", "str", "Tables") });
-
-		parsers->insert({ "get_table_selections", mvPythonParser({
-			{mvPythonDataType::String, "table"}
-		}, "Retrieves data from storage.", "List[List[int]]", "Tables") });
-	}
-
 	static void AddItemCommands(std::map<std::string, mvPythonParser>* parsers)
 	{
 		parsers->insert({ "get_windows", mvPythonParser({
@@ -883,274 +803,6 @@ namespace Marvel {
 		parsers->insert({ "show_source", mvPythonParser({
 			{mvPythonDataType::String, "file"},
 		}, "Shows the source code for a file.", "None", "Standard Windows") });
-	}
-
-	static void AddThemeCommands(std::map<std::string, mvPythonParser>* parsers)
-	{
-		parsers->insert({ "add_additional_font", mvPythonParser({
-			{mvPythonDataType::String, "file", "ttf or otf file"},
-			{mvPythonDataType::Optional},
-			{mvPythonDataType::Float, "size"},
-			{mvPythonDataType::String, "glyph_ranges", "options: korean, japanese, chinese_full, chinese_simplified_common, cryillic, thai, vietnamese"},
-			{mvPythonDataType::KeywordOnly},
-			{mvPythonDataType::IntList, "custom_glyph_chars", ""},
-			{mvPythonDataType::Object, "custom_glyph_ranges", ""},
-		}, "Adds additional font.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_theme", mvPythonParser({
-			{mvPythonDataType::String, "theme"}
-		}, "Set the application's theme to a built-in theme.", "None", "Themes and Styles") });
-
-		parsers->insert({ "get_theme_item", mvPythonParser({
-			{mvPythonDataType::Integer, "item"}
-		}, "Returns a theme item's color", "(float, float, float, float)", "Themes and Styles") });
-
-		parsers->insert({ "get_theme", mvPythonParser({
-		}, "Returns the current theme.", "str", "Themes and Styles") });
-
-		parsers->insert({ "set_item_color", mvPythonParser({
-			{mvPythonDataType::String, "item"},
-			{mvPythonDataType::Integer, "style"},
-			{mvPythonDataType::FloatList, "color"}
-		}, "Sets an color style for a single item.", "None", "Themes and Styles") });
-
-		parsers->insert({ "clear_item_color", mvPythonParser({
-			{mvPythonDataType::String, "item"},
-		}, "Clears individual color styles for an item.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_theme_item", mvPythonParser({
-			{mvPythonDataType::Integer, "item"},
-			{mvPythonDataType::Integer, "r"},
-			{mvPythonDataType::Integer, "g"},
-			{mvPythonDataType::Integer, "b"},
-			{mvPythonDataType::Integer, "a"}
-		}, "Sets a theme item.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_window_padding", mvPythonParser({
-			{mvPythonDataType::Float, "x"},
-			{mvPythonDataType::Float, "y"},
-		}, "Sets padding within a window.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_frame_padding", mvPythonParser({
-			{mvPythonDataType::Float, "x"},
-			{mvPythonDataType::Float, "y"},
-		}, "Sets padding within a framed rectangle (used by most widgets).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_item_spacing", mvPythonParser({
-			{mvPythonDataType::Float, "x"},
-			{mvPythonDataType::Float, "y"},
-		}, "Sets horizontal and vertical spacing between widgets/lines.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_item_inner_spacing", mvPythonParser({
-			{mvPythonDataType::Float, "x"},
-			{mvPythonDataType::Float, "y"},
-		}, "Sets horizontal and vertical spacing between within elements of a composed widget (e.g. a slider and its label).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_touch_extra_padding", mvPythonParser({
-			{mvPythonDataType::Float, "x"},
-			{mvPythonDataType::Float, "y"},
-		}, "Expand reactive bounding box for touch-based system where touch position is not accurate enough. Unfortunately we don't sort widgets so priority on overlap will always be given to the first widget. So don't grow this too much!", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_indent_spacing", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets horizontal indentation when e.g. entering a tree node. Generally == (FontSize + FramePadding.x*2).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_scrollbar_size", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets width of the vertical scrollbar, Height of the horizontal scrollbar.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_grab_min_size", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets minimum width/height of a grab box for slider/scrollbar.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_window_border_size", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets thickness of border around windows. Generally set to 0.0 or 1.0. (Other values are not well tested and more CPU/GPU costly).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_child_border_size", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets thickness of border around child windows. Generally set to 0.0 or 1.0. (Other values are not well tested and more CPU/GPU costly).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_popup_border_size", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets thickness of border around popup/tooltip windows. Generally set to 0.0 or 1.0. (Other values are not well tested and more CPU/GPU costly).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_frame_border_size", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets thickness of border around frames. Generally set to 0.0 or 1.0. (Other values are not well tested and more CPU/GPU costly).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_tab_border_size", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets thickness of border around tabs.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_window_rounding", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets Radius of window corners rounding. Set to 0.0fto have rectangular windows. Large values tend to lead to variety of artifacts and are not recommended.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_child_rounding", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets radius of child window corners rounding. Set to 0.0 to have rectangular windows.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_frame_rounding", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets radius of frame corners rounding. Set to 0.0 to have rectangular frame (used by most widgets).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_popup_rounding", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets radius of popup window corners rounding. (Note that tooltip windows use WindowRounding).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_scrollbar_rounding", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets radius of grab corners for scrollbar.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_grab_rounding", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets radius of grabs corners rounding. Set to 0.0 to have rectangular slider grabs.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_tab_rounding", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets radius of upper corners of a tab. Set to 0.0 to have rectangular tabs.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_window_title_align", mvPythonParser({
-			{mvPythonDataType::Float, "x"},
-			{mvPythonDataType::Float, "y"},
-		}, "Sets alignment for title bar text. Defaults to (0.0,0.5) for left-aligned,vertically centered.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_window_menu_button_position", mvPythonParser({
-			{mvPythonDataType::Integer, "value"}
-		}, "Sets side of the collapsing/docking button in the title bar (None/Left/Right). Defaults to mvGuiDir_Left.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_color_button_position", mvPythonParser({
-			{mvPythonDataType::Integer, "value"}
-		}, "Sets side of the color button in the ColorEdit4 widget (left/right). Defaults to mvGuiDir_Right.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_button_text_align", mvPythonParser({
-			{mvPythonDataType::Float, "x"},
-			{mvPythonDataType::Float, "y"},
-		}, "Sets alignment of button text when button is larger than text. Defaults to (0.5, 0.5) (centered).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_selectable_text_align", mvPythonParser({
-			{mvPythonDataType::Float, "x"},
-			{mvPythonDataType::Float, "y"},
-		}, "Sets alignment of selectable text. Defaults to (0.0, 0.0) (top-left aligned). It's generally important to keep this left-aligned if you want to lay multiple items on a same line.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_display_safe_area_padding", mvPythonParser({
-			{mvPythonDataType::Float, "x"},
-			{mvPythonDataType::Float, "y"},
-		}, "Sets if you cannot see the edges of your screen (e.g. on a TV) increase the safe area padding. Apply to popups/tooltips as well regular windows. NB: Prefer configuring your TV sets correctly!", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_global_alpha", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets global alpha applies to everything in Dear PyGui.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_antialiased_lines", mvPythonParser({
-			{mvPythonDataType::Bool, "value"}
-		}, "Sets anti-aliasing on lines/borders. Disable if you are really tight on CPU/GPU.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_antialiased_fill", mvPythonParser({
-			{mvPythonDataType::Bool, "value"}
-		}, "Sets anti-aliasing on filled shapes (rounded rectangles, circles, etc.).", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_curve_tessellation_tolerance", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets Tessellation tolerance.", "None", "Themes and Styles") });
-
-		parsers->insert({ "set_style_circle_segment_max_error", mvPythonParser({
-			{mvPythonDataType::Float, "value"}
-		}, "Sets maximum error (in pixels) allowed when using draw_circle()or drawing rounded corner rectangles with no explicit segment count specified. Decrease for higher quality but more geometry.", "None", "Themes and Styles") });
-
-		parsers->insert({ "get_style_window_padding", mvPythonParser({
-		}, "Gets padding within a window.", "List[float]", "Themes and Styles") });
-
-		parsers->insert({ "get_style_frame_padding", mvPythonParser({
-		}, "Gets padding within a framed rectangle (used by most widgets).", "List[float]", "Themes and Styles") });
-
-		parsers->insert({ "get_style_item_spacing", mvPythonParser({
-		}, "Gets horizontal and vertical spacing between widgets/lines.", "List[float]", "Themes and Styles") });
-
-		parsers->insert({ "get_style_item_inner_spacing", mvPythonParser({
-		}, "Gets horizontal and vertical spacing between within elements of a composed widget (e.g. a slider and its label).", "List[float]", "Themes and Styles") });
-
-		parsers->insert({ "get_style_touch_extra_padding", mvPythonParser({
-		}, "Get touch extra padding.", "List[float]", "Themes and Styles") });
-
-		parsers->insert({ "get_style_indent_spacing", mvPythonParser({
-		}, "Gets horizontal indentation when e.g. entering a tree node. Generally == (FontSize + FramePadding.x*2).", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_scrollbar_size", mvPythonParser({
-		}, "Gets width of the vertical scrollbar, Height of the horizontal scrollbar.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_grab_min_size", mvPythonParser({
-		}, "Gets minimum width/height of a grab box for slider/scrollbar.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_window_border_size", mvPythonParser({
-		}, "Gets thickness of border around windows.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_child_border_size", mvPythonParser({
-		}, "Gets thickness of border around child windows.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_popup_border_size", mvPythonParser({
-		}, "Gets thickness of border around popup/tooltip windows.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_frame_border_size", mvPythonParser({
-		}, "Gets thickness of border around frames.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_tab_border_size", mvPythonParser({
-		}, "Gets thickness of border around tabs.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_window_rounding", mvPythonParser({
-		}, "Gets radius of window corners rounding.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_child_rounding", mvPythonParser({
-		}, "Gets radius of child window corners rounding.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_frame_rounding", mvPythonParser({
-		}, "Gets radius of frame corners rounding.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_popup_rounding", mvPythonParser({
-		}, "Gets radius of popup window corners rounding. (Note that tooltip windows use WindowRounding).", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_scrollbar_rounding", mvPythonParser({
-		}, "Gets radius of grab corners for scrollbar.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_grab_rounding", mvPythonParser({
-		}, "Gets radius of grabs corners rounding. Set to 0.0 to have rectangular slider grabs.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_tab_rounding", mvPythonParser({
-		}, "Gets radius of upper corners of a tab. Set to 0.0 to have rectangular tabs.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_window_title_align", mvPythonParser({
-		}, "Gets alignment for title bar text. Defaults to (0.0,0.5) for left-aligned,vertically centered.", "List[float]", "Themes and Styles") });
-
-		parsers->insert({ "get_style_window_menu_button_position", mvPythonParser({
-		}, "Gets side of the collapsing/docking button in the title bar (None/Left/Right). Defaults to mvGuiDir_Left.", "int", "Themes and Styles") });
-
-		parsers->insert({ "get_style_color_button_position", mvPythonParser({
-		}, "Gets side of the color button in the ColorEdit4 widget (left/right). Defaults to mvGuiDir_Right.", "int", "Themes and Styles") });
-
-		parsers->insert({ "get_style_button_text_align", mvPythonParser({
-		}, "Gets alignment of button text when button is larger than text. Defaults to (0.5, 0.5) (centered).", "List[float]", "Themes and Styles") });
-
-		parsers->insert({ "get_style_selectable_text_align", mvPythonParser({
-		}, "Gets alignment of selectable text. Defaults to (0.0, 0.0) (top-left aligned).", "List[float]", "Themes and Styles") });
-
-		parsers->insert({ "get_style_display_safe_area_padding", mvPythonParser({
-		}, "Gets safe area padding. Applies to popups/tooltips as well regular windows.", "List[float]", "Themes and Styles") });
-
-		parsers->insert({ "get_style_global_alpha", mvPythonParser({
-		}, "Gets global alpha applies to everything in Dear PyGui.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_antialiased_lines", mvPythonParser({
-		}, "Gets anti-aliasing on lines/borders.", "bool", "Themes and Styles") });
-
-		parsers->insert({ "get_style_antialiased_fill", mvPythonParser({
-		}, "Gets anti-aliasing on filled shapes (rounded rectangles, circles, etc.).", "bool", "Themes and Styles") });
-
-		parsers->insert({ "get_style_curve_tessellation_tolerance", mvPythonParser({
-		}, "Gets Tessellation tolerance.", "float", "Themes and Styles") });
-
-		parsers->insert({ "get_style_circle_segment_max_error", mvPythonParser({
-		}, "Gets maximum error (in pixels) allowed when using draw_circle()or drawing rounded corner rectangles with no explicit segment count specified.", "float", "Themes and Styles") });
 	}
 
 	static void AddMenuCommands(std::map<std::string, mvPythonParser>* parsers)
